==> 1.main.ipynb <==
# Importing Packages
import pandas as pd
from webdriver_manager.chrome import ChromeDriverManager
from selenium import webdriver
from datetime import date, timedelta, datetime
from IPython.core.display import clear_output
from random import randint
from requests import get
from selenium import webdriver
from selenium.webdriver.common.action_chains import ActionChains
from time import sleep
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.keys import Keys
from fake_useragent import UserAgent
from ipynb.fs.full.myfunction import get_info

final=pd.read_csv('final.csv')

final.shape

# URL of search
url = "https://www.linkedin.com/jobs/search/?f_PP=106356396%2C106812674%2C102253611%2C103969862&f_TPR=r86400&geoId=105646813&location=Spain&sortBy=R"

#Calling Function
data=get_info(url,no_jobs=600,seniority="Junior",industry="Stuff")

#Checking if no_jobs is correct
data.shape

data.head()

final=final.append(data, ignore_index=True)

final.shape

#Export data to CSV
final=final.to_csv('final.csv',index=False)
