==> 2.data-preanalysis.ipynb <==
#import libraries
import numpy as np
import pandas as pd

#read csv
df=pd.read_csv("final.csv")

df.shape

df.columns

#Delete some columns
df=df.drop(columns=["Company Industry", "Job Seniorty", "Job link"])

df.columns

#delete %100 duplicated rows
df=df.drop_duplicates() 

df.shape

#formating columns to lower
df.columns=df.columns.str.lower()

df.columns

#Checking data types
df.dtypes

#Coverting date column from str to date
df['date'] = pd.to_datetime(df['date'])

#Checking data types
df.dtypes

#Checking nulls
df.isnull().sum()

#Filling Null Values
df["location"]=df["location"].fillna("0")

#Checking nulls
df.isnull().sum()

#Checking company columnn
df.company.value_counts()

#Checking title columnn
df.title.value_counts()

#Checking location columnn
df.location.value_counts()

#Export data to CSV
df=df.to_csv('df.csv',encoding='utf-8-sig', index=False)
