==> 3.data-location-cleaning.ipynb <==
#import libraries
import numpy as np
import pandas as pd
import csv

df=pd.read_csv("df.csv")

#Checking location columnn
df.location.value_counts()

#First cleaning
df.location=df.location.str.split("-").str[0]
df.location=df.location.str.split("/").str[0]
df.location=df.location.str.split("(").str[0]
df.location=df.location.str.split(",").str[0]
df.location=df.location.str.split(" y alrededores").str[0]
df.location=df.location.str.split(" y alrededore").str[0]
df.location=df.location.str.lstrip()
df.location=df.location.str.rstrip()

#Second cleaning
df.loc[df["location"].str.contains("Madrid", case=False), "location"] = "Madrid"
df.loc[df["location"].str.contains("Barcelona", case=False), "location"] = "Barcelona"


#Comunity Function
Andalucía=["Andalucía","Almería","Cádiz","Córdoba","Granada","Huelva","Jaén","Málaga","Sevilla"
          ,"Jerez de la Frontera","Marbella","Sierra de Cádiz","Dos Hermanas","Fuengirola"
          ,"Estepona","Mijas","Alcalá de Guadaíra","El Puerto de Santa María"
          ,"Torrevieja","Benalmádena","Torremolinos","Mairena del Aljarafe","El Ejido","Motril"
          ,"Linares","San Fernando","Tomares","Algeciras","Los Barrios","Vélez","Roquetas de Mar"
          ,"Arcos de la Frontera","Chiclana de la Frontera","Martos","Utrera"
          ,"La Línea de la Concepción","Antequera","Armilla","Rota","Sanlúcar de Barrameda"
          ,"Adra","Isla Cristina","Puerto Real","La Rinconada","Lepe","Alcalá La Real","Andújar"
          ,"Ronda","Montilla","Mojácar","Finestrat","Catarroja"
           ,"Zahara de los Atunes","Barbate","Úbeda","Puente Genil","Petrer","Coria del Río"
          ,"Coín","Baza","Palma del Río","Nerja","Nerja","Maracena","Aljaraque","Almuñécar"
          ,"Los Palacios y Villafranca","Alhaurín de la Torre","Aguadulce","Alcaudete","Baeza"
          ,"Conil de la Frontera","El Viso del Alcoranda","El Viso del Alcor"]

Aragón=["Aragón","Huesca","Teruel","Zaragoza","Llanes","Zuera","Alcañiz","Calatayud","Utebo"
        ,"Tarazona de la Mancha","Caspe", "Andorra","Monzón","Navia","Barbastro"
        ,"Ejea de los Caballeros"]

Asturias =["Asturias","Avilés","Gijón","Mieres","Oviedo","Siero","Langreo","Castrillón"
           ,"Ribadesella","Llanera","Pola de Siero","Cangas de Onís","Principado de Asturias"
          ,"Cudillero","Corvera de Asturias","Villaviciosa","Tineo","Parres","Nava","Grado"
          ,"Cangas del Narcea"]

Baleares=["Acampada Menorca","Acampada Palma de Mallorca","Palma de Mallorca"
          ,"Eivissa","Manacor","Es Mercadal","Islas Baleares","Ciutadella de Menorca","Marratxí"
         ,"Calvià","Formentera","Cala Millor","Maó"]

Canarias =["Palmas" ,"Santa Cruz de Tenerife","Fuerteventura","Las Palmas","Lanzarote","Arona"
          ,"Las Palmas de Gran Canaria","Palma","San Cristóbal de la Laguna","Arrecife"
          ,"Telde","Puerto del Rosario","Adeje","Corralejo","Yaiza","Puerto de la Cruz"
          ,"Santa Lucía de Tirajana","Valsequillo de Gran Canaria","Teguise"
          ,"Tacoronte","San Bartolomé de Tirajana","Maspalomas","El Pinar del Hierro"
          ,"Tías","La Palma","San Sebastián de la Gomera","Agüimes","Breña Baja","Candelaria"]

Cantabria=["Santander","Torrelavega","Castro-Urdiales","Camargo","Autonomía de Cantabria"
          ,"Cieza","Ampuero","Anievas","Piélagos","El Astillero","Santa Cruz de Bezana","Laredo"
          ,"Santoña","Los Corrales de Buelna","Reinosa","Santa María de Cayón","Suances"
           ,"Cabezón de la Sal","Reocín","Colindres","Medio Cudeyo","Castro","Noja"
          ,"Soto del Barco","Santillana del Mar","San Vicente de la Barquera"]


Castilla_Mancha=["Albacete", "Ciudad Real", "Cuenca", "Guadalajara", "Toledo","García"
                ,"Talavera de la Reina","Illescas","Azuqueca de Henares","Casarrubios del Monte"
                ,"Seseña","Cabanillas del Campo","Torija","Valdepeñas","Puertollano"
                ,"Tomelloso","Tarancón","Yeles","Alcázar de San Juan","Alcázar","Almansa"
                ,"Alovera","Hellín","La Roda"]


Castilla_León =["Ávila", "Burgos", "León", "Salamanca", "Segovia", "Soria", "Valladolid", "Zamora"
               ,"Palencia","Castilla y León","Región de León","Aranda de Duero","Miranda de Ebro"
               ,"Coruña","Coruña del Conde","Ponferrada","Viana de Duero"
               ,"Medina del Campo","Benavente","Puebla de Sanabria"
                ,"Tabladillo","Aguilar de Campoo","Laguna de Duero","Astorga"]


Cataluña=["Cataluña","Barcelona","Tarragona","Lérida","Gerona","Hospitalet de Llobregat","Badalona","Tarrasa"
         ,"Sabadell","Mataró","Sant Cugat del Vallès","Terrassa","El Prat de Llobregat"
         ,"Vic","Mollet del Vallès","Cornellà de Llobregat","Granollers","Barberà del Vallès"
         ,"Manresa","Sant Adrià de Besòs","Sant Boi de Llobrega","Reus","Vilafranca del Penedès"
         ,"Vilanova i la Geltrù","Sant Pere de Ribes","Reus","Rubí","Martorell"
         ,"Montornès del Vallès","Olesa de Montserrat","Castelldefels","Santa Coloma de Gramenet"
         ,"Figueres","Igualada","La Roca del Vallès","Olot","Viladecans","Sant Quirze del Vallès"
         ,"Castellar del Vallès","Lloret de Mar","Sant Boi de Llobregat","la Roca del Vallès"
          ,"Molins de Rei","Sitges","Cerdanyola del Vallès","Esplugues de Llobregat"
         ,"Gavà","Parets del Vallès","Sant Andreu de la Barca","Santa Perpetua de Mogoda"
         ,"Cambrils","Montcada I Reixac","Sant Felíu de Llobregat","Tordera","Pals","Salou"
          ,"Vilamalla","Blanes","Caldes de Montbui","la Selva del Camp","Lliçà d'Amunt"
          ,"Lliçà de Vall","Sant Esteve Sesrovires","Sant Joan Despí","Abrera"
         ,"Castellbisbal","El Vendrell","Sant Sadurni d'Anoia","Calafell","Cervera","San Juan Despí"
         ,"Sant Vicenç dels Horts","Santa Oliva","Vilanova del Vallès","Tossa de Mar"
         ,"Tortosa","Ripollet","Premià de Mar","Palau de Plegamans","Palafrugell"
         ,"Llinars del Vallès","Les Franqueses del Vallès","Hostalric","Vilassar de Mar"
         ,"Riudellots de la Selva","Castell d'Aro","Sant Fost de Campsentelles"
         ,"Piera","El Masnou","Cubelles","Castelló d'Empúries","Tàrrega","Tàrrega"
         ,"Santa Maria de Palautordera","Sant Jaume dels Domenys","Sant Fruitós de Bages"
         ,"Palamós","Mollerussa","Balaguer","Cervelló","Sant Andreu de Llavaneres","Sant Celoni"
         ,"Sant Cugat Sesgarrigues","Sant Just Desvern","Sant Vicenç de Castellet"
         ,"Santa Margarida i els Monjos","Montserrat","Banyoles","Calella"]

Extremadura =["Badajoz", "Cáceres","Extremadura","Mérida","Plasencia","Almendralejo"
             ,"Don Benito","Zafra","Villanueva de la Serena"]

Galicia=["La Coruña", "Lugo", "Orense", "Pontevedra","Vigo","A Rigueira"
        ,"Santiago de Compostela","Galicia ","Noia","San Cibrao das Viñas","Vilagarcía de Arousa"
        ,"Ferrol","Tarrio","Boiro","O Porriño","Gándara","Santa Cristina de Paradela"
        ,"Ribadeo","Bueu","Arteixo","Oleiros","O Toural","Monforte de Lemos","O Grove","Ames"
        ,"Cangas"]

Rioja=["Alfaro","Arnedo","Calahorra","Haro","Lardero","Logroño","Nájera","Santo Domingo de la Calzada"
,"Villamediana de Iregu","Rioja","Torrecilla en Cameros"]

Madrid =["Alcalá de Henares","Alcobendas","Alcorcón","Aranjuez","Arganda del Rey"
         ,"Collado Villalba","Colmenar Viejo","Coslada","Fuenlabrada","Getafe","Leganés"
    ,"Madrid","Majadahonda","Móstoles","Navalcarnero","Parla","San Lorenzo de El Escorial"
    ,"Torrejón de Ardoz","Torrelaguna","Valdemoro","Getafe","Pozuelo de Alarcón","Majadahonda"
        ,"Tres Cantos","Pinto","San Sebastián de los Reyes","San Fernando de Henares"
        ,"Rivas","Boadilla del Monte","Ciempozuelos","Loeches","Algete"
        ,"San Agustín del Guadalix","Paracuellos de Jarama","Guadarrama"
        ,"Villaviciosa de Odón","Arroyomolinos","Daganzo de Arriba","Villanueva de la Cañada"
        ,"Torrelodones","Antigua","Mejorada del Campo","Centro","San Lorenzo del Escorial"
        ,"Camarma de Esteruelas","Navacerrada","Barajas","Villalbilla","San Martín de Valdeiglesias"
        ,"Las Tablas","El Escorial","Villalba","Torrejón de Velasco"
         ,"Meco","Griñón","Ajalvir","Fuente el Saz de Jarama","Galapagar"]




Murcia =["Murcia","Cartagena","Lorca","Molina de Segura","Alcantarilla","Torre Pacheco"
        ,"Alhama de Murcia","Lorquí","Águilas","Cabezo de Torres","Puerto de Mazarrón"
        ,"Yecla","San Javier","Mazarrón","Totana","Caravaca de la Cruz","Jumilla"
        ,"San Pedro del Pinatar","Fuente Álamo de Murcia","Región de Murcia","La Unión","Aljucer"
        ,"Las Torres de Cotillas","Lo Gea"]

Navarra=["Pamplona","Tudela","Gaztelu","Comunidad Foral de Navarra","Estella","Sestao"
         ,"Viana","Sarriguren","Burlada"]

País_Vasco =["Bilbao", "San Sebastián", "Vitoria","Álava","Guipúzcoa","Vizcaya","Donostia"
            ,"Alava","País Vasco","San Vicente de Barakaldo","Irun","Oñati","Amorebieta"
            ,"Santurtzi","Basauri","Portugalete","Tolosa","Beasain","Getxo"
            ,"Lezo","Rentería","Zarauz","Mondragón","Llodio","Erandio","Berango"
             ,"Oiartzun","Abadiño","Algorta","La Manga del Mar Menor","Iurreta","Ricla","Okondo"
            ,"Azpeitia","Galdakao","Hernani"]




Valencia=["Comunidad Valenciana","Valencia","Alicante","Castellón","Denia","Jávea","Nules","Alcoy"
         ,"Valencia de Alcántara","Castellón de la Plana","Elche","Paterna","Dénia"
         ,"Benidorm","Gandia","Santa Pola","Torrent","Pau","Onda","Requena","Sagunto","El Palmar"
         ,"Ontinyent","l'Alcora","Orihuela","Valls","Alzira","Xàtiva","Lucena","Vinaròs","Altea"
         ,"Villanueva de Castellón","Formentera del Segura","Quart de Poblet"
         ,"San Vicente del Raspeig","Aldaia","Albuixech","Burjassot","Calpe","Llíria"
         ,"Manises","Mislata","Benicasim","Serra","Sant Joan d'Alacant","Sueca"
         ,"La Pobla de Vallbona","Alcúdia","Benicarló","Crevillente","Cullera","Elda"
          ,"Gandía","San Bartolomé","Xirivella","Vallés","Vall de Uxó"
         ,"Oliva","Almazora","Villajoyosa","Silla","Oropesa del Mar","Puçol","Ondara"
         ,"Guardamar del Segura","Cheste","San Antonio de Benagéber","Picassent","L'Alfàs del Pi"
         ,"Bétera","Borriana","El Campello","Masías Blancas","Massamagrell","Alboraya"
         ,"Alfafar","Almussafes"]

Remote= ["España","0","Arteaga","Nueva York","Pau","Vila","Palau","Lon","Campos","Riba"
        ,"Camas","Rus","Vitre","Roses","Quer"]

Cueta_Melilla=["Melilla","Ceuta"]


#Comunity function
def comunidad(x):
    if x in Andalucía :
        return "Andalucía"
    elif x in Aragón :
        return "Aragón"
    elif x in Asturias :
        return "Asturias"
    elif x in Baleares:
        return "Baleares"
    elif x in Canarias :
        return "Canarias"
    elif x in Cantabria:
        return "Cantabria"
    elif x in Castilla_Mancha:
        return "Castilla_Mancha"
    elif x in Castilla_León:
        return "Castilla_León"
    elif x in Cataluña:
        return "Cataluña"
    elif x in Extremadura :
        return "Extremadura"
    elif x in Galicia:
        return "Galicia"
    elif x in Rioja:
        return "Rioja"
    elif x in Madrid :
        return "Madrid"
    elif x in Murcia :
        return "Murcia"
    elif x in Navarra :
        return "Navarra"
    elif x in País_Vasco:
        return "País_Vasco"
    elif x in Valencia :
        return "Valencia"
    elif x in Cueta_Melilla :
        return "Cueta_Melilla"
    elif x in Remote :
        return "Remote"
    else:
        return "Other"

#Comunity Mapping
df["location"] =df["location"].map(comunidad)

#Checking location columnn
df.location.value_counts()

#Export data to CSV
df=df.to_csv('df2.csv',encoding='utf-8-sig', index=False)
