==> 4-data-title-cleaning.ipynb <==
#import libraries
import numpy as np
import pandas as pd
import csv

df=pd.read_csv("df2.csv")

#Checking title columnn
df.title.value_counts()

#First Cleaning
#df.title=df.title.str.split("Senior").str[0]
df.title=df.title.str.lstrip()
df.title=df.title.str.rstrip()
df.title=df.title.str.lower()

#TECH

#Frontend
df.loc[df["title"].str.contains("front", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("wordpress", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("php", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("xamarin", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("angular", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("node.js", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("kotlin", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("shopify", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("aem developer", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("golang", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("drupal", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("go developer", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("react", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("ruby", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("swift", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("vue.js", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("rpa developer", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("html", case=False), "title"] = "Frontend_Developer"
df.loc[df["title"].str.contains("devops", case=False), "title"] = "Frontend_Developer"

#Backend
df.loc[df["title"].str.contains("back", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("nodejs", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("magento", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("integration", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("api developer", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("laravel", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("liferay", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("solidity", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("unity", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("abap", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("apigee", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("stack", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("web developer", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("desarrollador", case=False), "title"] = "Backend_Developer"
df.loc[df["title"].str.contains("hardware", case=False), "title"] = "Backend_Developer"

#Data
df.loc[df["title"].str.contains("data", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("datos", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("machine", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("intelligence", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("inteligencia", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("tableau", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("analyst", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("bi engineer", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("analysis", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("analytics", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("business consultant", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("analista", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("power bi", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("scraping", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("bi developer", case=False), "title"] = "Data_Analytics"
df.loc[df["title"].str.contains("analytical", case=False), "title"] = "Data_Analytics"
 

#Security
df.loc[df["title"].str.contains("ciberseguridad", case=False), "title"] = "Cybersecurity"
df.loc[df["title"].str.contains("cybersecurity", case=False), "title"] = "Cybersecurity"
df.loc[df["title"].str.contains("cyber", case=False), "title"] = "Cybersecurity"
df.loc[df["title"].str.contains("security", case=False), "title"] = "Cybersecurity"
df.loc[df["title"].str.contains("seguridad", case=False), "title"] = "Cybersecurity"
df.loc[df["title"].str.contains("virtualización", case=False), "title"] = "Cybersecurity"



#Cloud
df.loc[df["title"].str.contains("cloud", case=False), "title"] = "Cloud_Expert"
df.loc[df["title"].str.contains("azure", case=False), "title"] = "Cloud_Expert"
df.loc[df["title"].str.contains("microsoft", case=False), "title"] = "Cloud_Expert"
df.loc[df["title"].str.contains("aws", case=False), "title"] = "Cloud_Expert"
df.loc[df["title"].str.contains("google", case=False), "title"] = "Cloud_Expert"
df.loc[df["title"].str.contains("Salesforce", case=False), "title"] = "Cloud_Expert"
df.loc[df["title"].str.contains("sales force consultant", case=False), "title"] = "Cloud_Expert"
df.loc[df["title"].str.contains("crm", case=False), "title"] = "Cloud_Expert"


#UX/UI
df.loc[df["title"].str.contains("ux/ui", case=False), "title"] = "UX/UI_Designer"
df.loc[df["title"].str.contains("designer", case=False), "title"] = "UX/UI_Designer"
df.loc[df["title"].str.contains("diseñador", case=False), "title"] = "UX/UI_Designer"
df.loc[df["title"].str.contains("ui ux web design", case=False), "title"] = "UX/UI_Designer"
df.loc[df["title"].str.contains("diseño", case=False), "title"] = "UX/UI_Designer"
df.loc[df["title"].str.contains("ui developer", case=False), "title"] = "UX/UI_Designer"
df.loc[df["title"].str.contains("ecommerce", case=False), "title"] = "UX/UI_Designer"
df.loc[df["title"].str.contains("e-commerce", case=False), "title"] = "UX/UI_Designer"
df.loc[df["title"].str.contains("comercio", case=False), "title"] = "UX/UI_Designer"
 
#System Admin
df.loc[df["title"].str.contains("arquitecto", case=False), "title"] = "System_Admin"
df.loc[df["title"].str.contains("administrator", case=False), "title"] = "System_Admin"
df.loc[df["title"].str.contains("database", case=False), "title"] = "System_Admin"
df.loc[df["title"].str.contains("system engineer", case=False), "title"] = "System_Admin"
df.loc[df["title"].str.contains("sistemas", case=False), "title"] = "System_Admin"
df.loc[df["title"].str.contains("systems", case=False), "title"] = "System_Admin"
df.loc[df["title"].str.contains("system", case=False), "title"] = "System_Admin"
df.loc[df["title"].str.contains("linux", case=False), "title"] = "System_Admin"
df.loc[df["title"].str.contains("windows", case=False), "title"] = "System_Admin"
df.loc[df["title"].str.contains("office", case=False), "title"] = "System_Admin"
df.loc[df["title"].str.contains("platform", case=False), "title"] = "System_Admin"
df.loc[df["title"].str.contains("platforma", case=False), "title"] = "System_Admin"


#SW_App Developer
df.loc[df["title"].str.contains("java", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("software", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("python", case=False), "title"] = "SW_APP_Developer" 
df.loc[df["title"].str.contains(".net", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("net", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("programador", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("automation", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("automatización", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("developer senior", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("junior developer", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("sw", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("microservices", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("android", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("aplicaciones", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("ios developer", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("application", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("mobile", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("app", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("ios engineer", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("flutter", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("movilidad", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("smart", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("smartphone", case=False), "title"] = "SW_APP_Developer"


#Network_Expert
df.loc[df["title"].str.contains("network", case=False), "title"] = "Network_Expert"
df.loc[df["title"].str.contains("telecomunicaciones", case=False), "title"] = "Network_Expert"
df.loc[df["title"].str.contains("telecommunications", case=False), "title"] = "Network_Expert"
df.loc[df["title"].str.contains("redes", case=False), "title"] = "Network_Expert"


#QA Engineer
df.loc[df["title"].str.contains("qa", case=False), "title"] = "QA_Engineer"
df.loc[df["title"].str.contains("test", case=False), "title"] = "QA_Engineer"
df.loc[df["title"].str.contains("quality assurance", case=False), "title"] = "QA_Engineer"
df.loc[df["title"].str.contains("software quality", case=False), "title"] = "QA_Engineer"
df.loc[df["title"].str.contains("tester", case=False), "title"] = "QA_Engineer"


#IT Support
df.loc[df["title"].str.contains("it manager", case=False), "title"] = "IT_Support"
df.loc[df["title"].str.contains("engineering", case=False), "title"] = "IT_Support"
df.loc[df["title"].str.contains("it support", case=False), "title"] = "IT_Support"
df.loc[df["title"].str.contains("it operational", case=False), "title"] = "IT_Support"
df.loc[df["title"].str.contains("technical support", case=False), "title"] = "IT_Support"
df.loc[df["title"].str.contains("internal", case=False), "title"] = "IT_Support"
df.loc[df["title"].str.contains("informatico", case=False), "title"] = "IT_Support"
df.loc[df["title"].str.contains("informática", case=False), "title"] = "IT_Support"
df.loc[df["title"].str.contains("informático", case=False), "title"] = "IT_Support"
df.loc[df["title"].str.contains("sharepoint", case=False), "title"] = "IT_Support"
df.loc[df["title"].str.contains("365", case=False), "title"] = "IT_Support" 

#Solution_Architect
df.loc[df["title"].str.contains("solution", case=False), "title"] = "Solution_Architect"
df.loc[df["title"].str.contains("solutions", case=False), "title"] = "Solution_Architect"
df.loc[df["title"].str.contains("architect", case=False), "title"] = "Solution_Architect"
df.loc[df["title"].str.contains("architecture", case=False), "title"] = "Solution_Architect"
df.loc[df["title"].str.contains("blockchain", case=False), "title"] = "Solution_Architect"
df.loc[df["title"].str.contains("tech lead", case=False), "title"] = "Solution_Architect"
df.loc[df["title"].str.contains("integración", case=False), "title"] = "Solution_Architect"
df.loc[df["title"].str.contains("integrador", case=False), "title"] = "Solution_Architect"
     

#MANAGMENT

#Director
df.loc[df["title"].str.contains("chief", case=False), "title"] = "Director"
df.loc[df["title"].str.contains("cfo", case=False), "title"] = "Director"
df.loc[df["title"].str.contains("directo", case=False), "title"] = "Director"
df.loc[df["title"].str.contains("ceo", case=False), "title"] = "Director"


#Project_Manager
df.loc[df["title"].str.contains("project", case=False), "title"] = "Project_Manager"
df.loc[df["title"].str.contains("program", case=False), "title"] = "Project_Manager"
df.loc[df["title"].str.contains("planificación", case=False), "title"] = "Project_Manager"
df.loc[df["title"].str.contains("proyectos", case=False), "title"] = "Project_Manager"
df.loc[df["title"].str.contains("proyectista", case=False), "title"] = "Project_Manager"

#Customer success
df.loc[df["title"].str.contains("success", case=False), "title"] = "CustomerSuccess_Manager"
df.loc[df["title"].str.contains("service manager", case=False), "title"] = "CustomerSuccess_Manager"
df.loc[df["title"].str.contains("excellence", case=False), "title"] = "CustomerSuccess_Manager"
df.loc[df["title"].str.contains("experience", case=False), "title"] = "CustomerSuccess_Manager"

#Product manager
df.loc[df["title"].str.contains("product", case=False), "title"] = "Product_Manager"

#Vendor Manager
df.loc[df["title"].str.contains("vendor", case=False), "title"] = "Vendor_manager"
df.loc[df["title"].str.contains("partner", case=False), "title"] = "Vendor_manager"
df.loc[df["title"].str.contains("compras", case=False), "title"] = "Vendor_manager"
df.loc[df["title"].str.contains("purchasing", case=False), "title"] = "Vendor_manager"
df.loc[df["title"].str.contains("comprador", case=False), "title"] = "Vendor_manager"
df.loc[df["title"].str.contains("compra", case=False), "title"] = "Vendor_manager"

#Area Manager
df.loc[df["title"].str.contains("jefe", case=False), "title"] = "Area_Manager"
df.loc[df["title"].str.contains("area", case=False), "title"] = "Area_Manager"
df.loc[df["title"].str.contains("gerente", case=False), "title"] = "Area_Manager"
df.loc[df["title"].str.contains("leader", case=False), "title"] = "Area_Manager"

#Quality & Process
#Scrum Master
df.loc[df["title"].str.contains("master", case=False), "title"] = "Scrum_Master"
df.loc[df["title"].str.contains("agile", case=False), "title"] = "Scrum_Master"
df.loc[df["title"].str.contains("digital transformation", case=False), "title"] = "Scrum_Master"

#Process_Expert
df.loc[df["title"].str.contains("process", case=False), "title"] = "Process_Expert"
df.loc[df["title"].str.contains("procesos", case=False), "title"] = "Process_Expert"
df.loc[df["title"].str.contains("proceso", case=False), "title"] = "Process_Expert"
df.loc[df["title"].str.contains("certification", case=False), "title"] = "Process_Expert"
df.loc[df["title"].str.contains("mejora", case=False), "title"] = "Process_Expert"
df.loc[df["title"].str.contains("continuous", case=False), "title"] = "Process_Expert"
 
#Quality Manager
df.loc[df["title"].str.contains("quality", case=False), "title"] = "Quality_Manager"
df.loc[df["title"].str.contains("calidad", case=False), "title"] = "Quality_Manager"

#SALES
#business developer
df.loc[df["title"].str.contains("development", case=False), "title"] = "Business_Developer"
df.loc[df["title"].str.contains("business", case=False), "title"] = "Business_Developer"
df.loc[df["title"].str.contains("negocio", case=False), "title"] = "Business_Developer"
df.loc[df["title"].str.contains("negocios", case=False), "title"] = "Business_Developer"

#Account Manager
df.loc[df["title"].str.contains("account", case=False), "title"] = "Account_Manager"
df.loc[df["title"].str.contains("cuentas", case=False), "title"] = "Account_Manager"

#Sales_Specialist
df.loc[df["title"].str.contains("sales", case=False), "title"] = "Sales_Specialist"
df.loc[df["title"].str.contains("comercial", case=False), "title"] = "Sales_Specialist"

#Presale_Specialist
df.loc[df["title"].str.contains("presale", case=False), "title"] = "Presale_Specialist"
df.loc[df["title"].str.contains("sales engineer", case=False), "title"] = "Presale_Specialist"
df.loc[df["title"].str.contains("bid", case=False), "title"] = "Presale_Specialist"

#Operations

#General Manager
df.loc[df["title"].str.contains("director", case=False), "title"] = "General_Manager"
df.loc[df["title"].str.contains("general manager", case=False), "title"] = "General_Manager"


#Regional Manager
df.loc[df["title"].str.contains("country", case=False), "title"] = "Regional_Manger"
df.loc[df["title"].str.contains("regional", case=False), "title"] = "Regional_Manger"
df.loc[df["title"].str.contains("b2b", case=False), "title"] = "Regional_Manger"


#Operations Staff
df.loc[df["title"].str.contains("delivery", case=False), "title"] = "Operations_Staff"
df.loc[df["title"].str.contains("operational", case=False), "title"] = "Operations_Staff"
df.loc[df["title"].str.contains("operacional", case=False), "title"] = "Operations_Staff"
df.loc[df["title"].str.contains("operations", case=False), "title"] = "Operations_Staff"
df.loc[df["title"].str.contains("operaciones", case=False), "title"] = "Operations_Staff"
df.loc[df["title"].str.contains("operation", case=False), "title"] = "Operations_Staff"

#Customer service
df.loc[df["title"].str.contains("customer service", case=False), "title"] = "Customer_Service"
df.loc[df["title"].str.contains("desk", case=False), "title"] = "Customer_Service"
df.loc[df["title"].str.contains("atención", case=False), "title"] = "Customer_Service"
df.loc[df["title"].str.contains("customer care", case=False), "title"] = "Customer_Service"
df.loc[df["title"].str.contains("call", case=False), "title"] = "Customer_Service"
df.loc[df["title"].str.contains("atencion", case=False), "title"] = "Customer_Service"
df.loc[df["title"].str.contains("contact", case=False), "title"] = "Customer_Service"
df.loc[df["title"].str.contains("soporte", case=False), "title"] = "Customer_Service"


#Field_Engineer
df.loc[df["title"].str.contains("reliability", case=False), "title"] = "Field_Engineer"
df.loc[df["title"].str.contains("ml engineer", case=False), "title"] = "Field_Engineer"
df.loc[df["title"].str.contains("ai/ml", case=False), "title"] = "Field_Engineer"
df.loc[df["title"].str.contains("field", case=False), "title"] = "Field_Engineer"

#Other_Engineer
df.loc[df["title"].str.contains("hardware engineer", case=False), "title"] = "Other_Engineer"
df.loc[df["title"].str.contains("electrical", case=False), "title"] = "Other_Engineer"
df.loc[df["title"].str.contains("ingeniero industrial", case=False), "title"] = "Other_Engineer"
df.loc[df["title"].str.contains("ingeniero/a industrial", case=False), "title"] = "Other_Engineer" 
df.loc[df["title"].str.contains("civil", case=False), "title"] = "Other_Engineer"

#Supervisor
df.loc[df["title"].str.contains("supervisor", case=False), "title"] = "Supervisor"
df.loc[df["title"].str.contains("coordinador", case=False), "title"] = "Supervisor"
df.loc[df["title"].str.contains("coordinador/a", case=False), "title"] = "Supervisor"

#Marketing & Avertising
df.loc[df["title"].str.contains("community manager", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("Marketing", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("social media", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("publicidad", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("growth", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("content", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("brand", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("planner", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("advertising", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("copywriter", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("marca", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("media", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("marketer", case=False), "title"] = "Marketing"
df.loc[df["title"].str.contains("marketplace", case=False), "title"] = "Marketing"


#Media
df.loc[df["title"].str.contains("redactor/a", case=False), "title"] = "Media"
df.loc[df["title"].str.contains("redactor", case=False), "title"] = "Media"
df.loc[df["title"].str.contains("periodista", case=False), "title"] = "Media"
df.loc[df["title"].str.contains("periodismo", case=False), "title"] = "Media"
df.loc[df["title"].str.contains("digital", case=False), "title"] = "Media"
 

#Finance & Banking

#Financial Advisor
df.loc[df["title"].str.contains("financial advisor", case=False), "title"] = "Financial_Advisor"
df.loc[df["title"].str.contains("financiero", case=False), "title"] = "Financial_Advisor"
df.loc[df["title"].str.contains("finanzas", case=False), "title"] = "Financial_Advisor"
df.loc[df["title"].str.contains("financiera", case=False), "title"] = "Financial_Advisor"
df.loc[df["title"].str.contains("financial", case=False), "title"] = "Financial_Advisor"
df.loc[df["title"].str.contains("startup", case=False), "title"] = "Financial_Advisor"


#Financial Controller
df.loc[df["title"].str.contains("controller", case=False), "title"] = "Financial_Controller"
df.loc[df["title"].str.contains("finance", case=False), "title"] = "Financial_Controller"
df.loc[df["title"].str.contains("contabilidad", case=False), "title"] = "Financial_Controller"
df.loc[df["title"].str.contains("financier@", case=False), "title"] = "Financial_Controller"
df.loc[df["title"].str.contains("contable", case=False), "title"] = "Financial_Controller"

#Lawyer
df.loc[df["title"].str.contains("abogado", case=False), "title"] = "Lawyer"
df.loc[df["title"].str.contains("abogada", case=False), "title"] = "Lawyer"
df.loc[df["title"].str.contains("contract", case=False), "title"] = "Lawyer"

#Consultant
df.loc[df["title"].str.contains("bank", case=False), "title"] = "Consultant"
df.loc[df["title"].str.contains("consultant", case=False), "title"] = "Consultant"
df.loc[df["title"].str.contains("consultor", case=False), "title"] = "Consultant"
df.loc[df["title"].str.contains("consulting", case=False), "title"] = "Consultant"
df.loc[df["title"].str.contains("banca", case=False), "title"] = "Consultant"
df.loc[df["title"].str.contains("branch", case=False), "title"] = "Consultant"

#Human Resources
df.loc[df["title"].str.contains("talent", case=False), "title"] = "Talent_Specialist"
df.loc[df["title"].str.contains("humanos", case=False), "title"] = "Talent_Specialist"
df.loc[df["title"].str.contains("recursos", case=False), "title"] = "Talent_Specialist"
df.loc[df["title"].str.contains("workforce", case=False), "title"] = "Talent_Specialist"
df.loc[df["title"].str.contains("hr", case=False), "title"] = "Talent_Specialist"
df.loc[df["title"].str.contains("rrhh", case=False), "title"] = "Talent_Specialist"
df.loc[df["title"].str.contains("recruiter", case=False), "title"] = "Talent_Specialist"
df.loc[df["title"].str.contains("hunter", case=False), "title"] = "Talent_Specialist"
df.loc[df["title"].str.contains("human", case=False), "title"] = "Talent_Specialist"

#Health Care
#Nurse
df.loc[df["title"].str.contains("enfermero/a", case=False), "title"] = "Nurse"
df.loc[df["title"].str.contains("nurse", case=False), "title"] = "Nurse"
df.loc[df["title"].str.contains("cuidadora", case=False), "title"] = "Nurse"
df.loc[df["title"].str.contains("enfermero", case=False), "title"] = "Nurse"
df.loc[df["title"].str.contains("enfermeros", case=False), "title"] = "Nurse"
df.loc[df["title"].str.contains("enfermería", case=False), "title"] = "Nurse"
df.loc[df["title"].str.contains("enfermeria", case=False), "title"] = "Nurse"
df.loc[df["title"].str.contains("enfermera", case=False), "title"] = "Nurse"
df.loc[df["title"].str.contains("enfermeras", case=False), "title"] = "Nurse"

#Doctor
df.loc[df["title"].str.contains("médico", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("medico", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("odontólogo", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("podologo", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("medicina", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("psicólogo", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("psicólogos", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("podólogo", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("patólogo", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("podólogos", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("psicologo/a", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("psicologo", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("logopeda", case=False), "title"] = "Doctor"
df.loc[df["title"].str.contains("odontologo", case=False), "title"] = "Doctor"

#Physiotherapist
df.loc[df["title"].str.contains("fisioterapeuta", case=False), "title"] = "Physiotherapist"
df.loc[df["title"].str.contains("fisioterapia", case=False), "title"] = "Physiotherapist"

#Pharmacy Technician
df.loc[df["title"].str.contains("farmacia", case=False), "title"] = "Pharmacy_Technician"
df.loc[df["title"].str.contains("pharmacist", case=False), "title"] = "Pharmacy_Technician"
df.loc[df["title"].str.contains("pharma", case=False), "title"] = "Pharmacy_Technician"
df.loc[df["title"].str.contains("farmaceutica", case=False), "title"] = "Pharmacy_Technician"
df.loc[df["title"].str.contains("farmacéutico", case=False), "title"] = "Pharmacy_Technician"
df.loc[df["title"].str.contains("farmaceutico", case=False), "title"] = "Pharmacy_Technician"
df.loc[df["title"].str.contains("farma", case=False), "title"] = "Pharmacy_Technician"
df.loc[df["title"].str.contains("farmacéuticos", case=False), "title"] = "Pharmacy_Technician"


#Health Asistant
df.loc[df["title"].str.contains("clinical", case=False), "title"] = "Health_Assitant"
df.loc[df["title"].str.contains("auxiliar de odontología", case=False), "title"] = "Health_Assitant"
df.loc[df["title"].str.contains("podología", case=False), "title"] = "Health_Assitant"
df.loc[df["title"].str.contains("hospital", case=False), "title"] = "Health_Assitant"
df.loc[df["title"].str.contains("salud", case=False), "title"] = "Health_Assitant"
df.loc[df["title"].str.contains("sanitario", case=False), "title"] = "Health_Assitant"

#Dentist
df.loc[df["title"].str.contains("dentist", case=False), "title"] = "Health_Dental"
df.loc[df["title"].str.contains("dentista", case=False), "title"] = "Health_Dental"
df.loc[df["title"].str.contains("dentistas", case=False), "title"] = "Health_Dental"
df.loc[df["title"].str.contains("dental", case=False), "title"] = "Health_Dental"

#Nutritionist
df.loc[df["title"].str.contains("nutricionista", case=False), "title"] = "Nutritionist"

#Veterinarian
df.loc[df["title"].str.contains("veterinario", case=False), "title"] = "Veterinarian"
df.loc[df["title"].str.contains("veterinaria", case=False), "title"] = "Veterinarian"

#Construction
df.loc[df["title"].str.contains("obra", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("fontanero", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("albañil", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("construir", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("construcción", case=False), "title"] = "Construction" 
df.loc[df["title"].str.contains("carpintero", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("electricista", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("aislamiento", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("carretillero", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("electromecánico", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("techo", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("reformas", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("renovación", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("fontanería", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("ducha", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("instalador", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("puertas", case=False), "title"] = "Construction"
df.loc[df["title"].str.contains("cristalero", case=False), "title"] = "Construction"

#Realestate
df.loc[df["title"].str.contains("inmobiliario", case=False), "title"] = "Realestate_Agent"
df.loc[df["title"].str.contains("inmobiliaria", case=False), "title"] = "Realestate_Agent"
df.loc[df["title"].str.contains("estate", case=False), "title"] = "Realestate_Agent"
df.loc[df["title"].str.contains("immobilier", case=False), "title"] = "Realestate_Agent"

#Education
df.loc[df["title"].str.contains("investigador", case=False), "title"] = "Researcher"
df.loc[df["title"].str.contains("researcher", case=False), "title"] = "Researcher"
df.loc[df["title"].str.contains("phd", case=False), "title"] = "Researcher"
df.loc[df["title"].str.contains("estudiante", case=False), "title"] = "Researcher"
df.loc[df["title"].str.contains("laboratorio", case=False), "title"] = "Researcher"
df.loc[df["title"].str.contains("research", case=False), "title"] = "Researcher"
df.loc[df["title"].str.contains("search", case=False), "title"] = "Researcher"

#Internship
df.loc[df["title"].str.contains("beca", case=False), "title"] = "Internship"

#Teacher
df.loc[df["title"].str.contains("maestros", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("maestro", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("professor", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("trainer", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("coaching", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("coach", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("teacher", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("tutor", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("educador/a", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("educador", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("education", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("educación", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("profesor", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("profesores", case=False), "title"] = "Teacher"
df.loc[df["title"].str.contains("profesor/a", case=False), "title"] = "Teacher"

#Services
df.loc[df["title"].str.contains("perros", case=False), "title"] = "Dog_Trainer"
df.loc[df["title"].str.contains("gato", case=False), "title"] = "Dog_Trainer"
df.loc[df["title"].str.contains("perro", case=False), "title"] = "Dog_Trainer"

#Services
df.loc[df["title"].str.contains("entrenador", case=False), "title"] = "Fitness"
df.loc[df["title"].str.contains("fitness", case=False), "title"] = "Fitness"
df.loc[df["title"].str.contains("yoga", case=False), "title"] = "Fitness"


df.loc[df["title"].str.contains("cocinero", case=False), "title"] = "Cook"
df.loc[df["title"].str.contains("chef", case=False), "title"] = "Cook"
df.loc[df["title"].str.contains("cociner@", case=False), "title"] = "Cook"
df.loc[df["title"].str.contains("cocinera", case=False), "title"] = "Cook"

df.loc[df["title"].str.contains("camarero", case=False), "title"] = "Waiter"
df.loc[df["title"].str.contains("cocina", case=False), "title"] = "Waiter"
df.loc[df["title"].str.contains("camarera", case=False), "title"] = "Waiter"
df.loc[df["title"].str.contains("camarera/o", case=False), "title"] = "Waiter"
df.loc[df["title"].str.contains("camarer@", case=False), "title"] = "Waiter"
df.loc[df["title"].str.contains("restaurante", case=False), "title"] = "Waiter"
df.loc[df["title"].str.contains("restaurant", case=False), "title"] = "Waiter"
 

df.loc[df["title"].str.contains("conductor", case=False), "title"] = "Driver"
df.loc[df["title"].str.contains("driver", case=False), "title"] = "Driver"
df.loc[df["title"].str.contains("transportista", case=False), "title"] = "Driver"
df.loc[df["title"].str.contains("transporte", case=False), "title"] = "Driver"


df.loc[df["title"].str.contains("limpiador", case=False), "title"] = "Housekeeper"
df.loc[df["title"].str.contains("limpieza", case=False), "title"] = "Housekeeper"
df.loc[df["title"].str.contains("limpia", case=False), "title"] = "Housekeeper"
df.loc[df["title"].str.contains("cleaning assistant", case=False), "title"] = "Housekeeper"
df.loc[df["title"].str.contains("limpeza", case=False), "title"] = "Housekeeper"
df.loc[df["title"].str.contains("doméstica", case=False), "title"] = "Housekeeper"


df.loc[df["title"].str.contains("operarios", case=False), "title"] = "Operator"
df.loc[df["title"].str.contains("teleoperador", case=False), "title"] = "Operator"
df.loc[df["title"].str.contains("agente", case=False), "title"] = "Operator"
df.loc[df["title"].str.contains("venta", case=False), "title"] = "Operator"
df.loc[df["title"].str.contains("operador", case=False), "title"] = "Operator"

df.loc[df["title"].str.contains("repartidor", case=False), "title"] = "Warehouse"
df.loc[df["title"].str.contains("distribuidor", case=False), "title"] = "Warehouse"
df.loc[df["title"].str.contains("almacen", case=False), "title"] = "Warehouse"
df.loc[df["title"].str.contains("almacén", case=False), "title"] = "Warehouse"
df.loc[df["title"].str.contains("almacén", case=False), "title"] = "Warehouse"

df.loc[df["title"].str.contains("azafato/a", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("azafatos/as", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("asistente", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("auxiliar", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("doméstico", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("hogar", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("cuidador", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("cuidado", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("cuidar", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("niñera", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("mudanza", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("domestico", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("canguros", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("canguro", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("domicilio", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("recepcionista", case=False), "title"] = "Auxiliary_Staff"
df.loc[df["title"].str.contains("jardinero", case=False), "title"] = "Auxiliary_Staff"



df.loc[df["title"].str.contains("cajero", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("cajera", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("cajas", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("vendedor", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("store manager", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("veterinarias", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("dependiente", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("dependienta", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("dependienta", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("dependent", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("dependient@s", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("tienda", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("decathlon", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("media markt", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("supermercado", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("carnicero", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("carniceria", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("carnicero/a", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("carnicera", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("vended@r", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("panadería", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("frutero/a", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("fruteria", case=False), "title"] = "Shop_Agent"
df.loc[df["title"].str.contains("fruta", case=False), "title"] = "Shop_Agent"



#Assitant
df.loc[df["title"].str.contains("secretaria", case=False), "title"] = "Assitant"
df.loc[df["title"].str.contains("personal assistant", case=False), "title"] = "Assitant"
df.loc[df["title"].str.contains("azafatos/as", case=False), "title"] = "Assitant"
df.loc[df["title"].str.contains("azafatos/as", case=False), "title"] = "PAssitant"
df.loc[df["title"].str.contains("relaciones", case=False), "title"] = "Assitant"


#Administrative
df.loc[df["title"].str.contains("administrativo", case=False), "title"] = "Administrative"
df.loc[df["title"].str.contains("gestor", case=False), "title"] = "Administrative"
df.loc[df["title"].str.contains("agestorías", case=False), "title"] = "Administrative"
df.loc[df["title"].str.contains("agestoría", case=False), "title"] = "Administrative"
df.loc[df["title"].str.contains("fisical", case=False), "title"] = "Administrative"
df.loc[df["title"].str.contains("asesor", case=False), "title"] = "Administrative"
df.loc[df["title"].str.contains("advisor", case=False), "title"] = "Administrative"


#Technician
df.loc[df["title"].str.contains("tecnico", case=False), "title"] = "Technician"
df.loc[df["title"].str.contains("operario", case=False), "title"] = "Technician"
df.loc[df["title"].str.contains("operator", case=False), "title"] = "Technician"
df.loc[df["title"].str.contains("fabrica", case=False), "title"] = "Technician"
df.loc[df["title"].str.contains("mantenimiento", case=False), "title"] = "Technician"
df.loc[df["title"].str.contains("operari", case=False), "title"] = "Technician"
df.loc[df["title"].str.contains(" técnico/a superior", case=False), "title"] = "Technician"

#Concierge
df.loc[df["title"].str.contains("conserje", case=False), "title"] = "Technician"
df.loc[df["title"].str.contains("conserjes", case=False), "title"] = "Technician"
df.loc[df["title"].str.contains("mecánico", case=False), "title"] = "Technician"
df.loc[df["title"].str.contains("taller", case=False), "title"] = "Technician"


#HairSalon
df.loc[df["title"].str.contains("peluquería", case=False), "title"] = "Welfare"
df.loc[df["title"].str.contains("peluquera", case=False), "title"] = "Welfare"
df.loc[df["title"].str.contains("peluquerO", case=False), "title"] = "Welfare"
df.loc[df["title"].str.contains("peluqueria", case=False), "title"] = "Welfare"

df.loc[df["title"].str.contains("masaje", case=False), "title"] = "Welfare"
df.loc[df["title"].str.contains("masajistas", case=False), "title"] = "Welfare"
df.loc[df["title"].str.contains("masajista", case=False), "title"] = "Welfare"


#Art & Design
df.loc[df["title"].str.contains("delineante", case=False), "title"] = "Art_Design"
df.loc[df["title"].str.contains("delineantes", case=False), "title"] = "Art_Design"
df.loc[df["title"].str.contains("artist", case=False), "title"] = "Art_Design"
df.loc[df["title"].str.contains("fotografía", case=False), "title"] = "Art_Design"
df.loc[df["title"].str.contains("fotógrafo", case=False), "title"] = "Art_Design"
df.loc[df["title"].str.contains("fotógrafos", case=False), "title"] = "Art_Design"
df.loc[df["title"].str.contains("arte", case=False), "title"] = "Art_Design"


# Checking title columnn
df.title.value_counts()

#Export data to CSV
df=df.to_csv('df3.csv',encoding='utf-8-sig', index=False)
