==> 5.Final Cleaning.ipynb <==
#import libraries
import numpy as np
import pandas as pd
import csv

df=pd.read_csv("df3.csv")

df.head()

df.title.value_counts()

#SW_Developer
df.loc[df["title"].str.contains("desarollador/a", case=False), "title"] = "SW_APP_Developer"
df.loc[df["title"].str.contains("desarollado", case=False), "title"] = "SW_APP_Developer"

#Marketing
df.loc[df["title"].str.contains("comunicación", case=False), "title"] = "Marketing"

#Network_Expert
df.loc[df["title"].str.contains("comunicaciones", case=False), "title"] = "Network_Expert"

#Other_Egineer
df.loc[df["title"].str.contains("ingeniero/a", case=False), "title"] = "Other_Engineer"
df.loc[df["title"].str.contains("ingeniería", case=False), "title"] = "Other_Engineer"


df.title.value_counts()

#Job Function
function={
"Frontend_Developer":"Tech","Backend_Developer":"Tech","Data_Analytics":"Tech"
,"Cybersecurity":"Tech","Cloud_Expert":"Tech","UX/UI_Designer":"Tech"
,"System_Admin":"Tech","SW_APP_Developer":"Tech","Network_Expert":"Tech"
,"QA_Engineer":"Tech","IT_Support":"Tech","Solution_Architect":"Tech"   
,"Director": "Management","Project_Manager":"Management","CustomerSuccess_Manager" :"Management"
,"Product_Manager" :"Management","Vendor_manager":"Management","Area_Manager":"Management"
,"Scrum_Master":"Quality & Process","Process_Expert":"Quality & Process"
,"Quality_Manager":"Quality & Process","Business_Developer":"Sales","Account_Manager":"Sales"
,"Sales_Specialist":"Sales","Presale_Specialist":"Sales"
,"General_Manager":"Operations","Regional_Manger":"Operations"
,"Operations_Staff":"Operations","Customer_Service":"Operations","Field_Engineer":"Operations"
,"Other_Engineer":"Operations","Supervisor":"Operations","Marketing":"Marketing","Media":"Marketing"
,"Financial_Advisor":"Finance","Financial_Controller":"Finance","Lawyer":"Finance"
,"Consultant":"Finance","Talent_Specialist":"Talent_Specialist","Nurse":"Health_Care"
,"Doctor":"Health_Care","Physiotherapist":"Health_Care"
,"Pharmacy_Technician":"Health_Care","Health_Assitant":"Health_Care"
,"Health_Dental":"Health_Care","nutritionist":"Health_Care","Veterinarian":"Health_Care"
,"Construction":"Construction","Realestate":"Realestate_Agent"
,"Researcher":"Education","Internship":"Education","Teacher":"Education"
,"Dog_Trainer":"Services","Fitness":"Services","Cook":"Services","Waiter":"Services"
,"Driver":"Services" ,"Housekeeper":"Services","Operator":"Services","Warehouse":"Services"
,"Auxiliary_Staff":"Services","Shop_Agent":"Services","Assitant":"Services"
,"Administrative":"Services","Technician":"Services","Welfare":"Services"
,"Art_Design":"Art_Design"    
}

#Creating Job Function Column
df['Job_function']=df['title'].map(function)
df['Job_function']=df['Job_function'].fillna("other")

df.head()

#Export data to CSV
df=df.to_csv('df4.csv',encoding='utf-8-sig', index=False)
