==> 7.data-data-cleaning.ipynb <==
#import libraries
import numpy as np
import pandas as pd

#read csv
df=pd.read_csv("data.csv")

df.shape

df.columns

#delete %100 duplicated rows
df=df.drop_duplicates() 

df.shape

#formating columns to lower
df.columns=df.columns.str.lower()

#Checking title columnn
df.title.value_counts()

df.title=df.title.str.lstrip()
df.title=df.title.str.rstrip()
df.title=df.title.str.lower()

df.loc[df["title"].str.contains("data analyst", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("analista", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("product", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("solutions", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("intelligence", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("analytics", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("research analyst", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("big data", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("consultant", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("strategy", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("bi analyst", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("quality", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("digital", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("marketing", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("senior analyst", case=False), "title"] = "Data_Analyst"
df.loc[df["title"].str.contains("technical analyst", case=False), "title"] = "Data_Analyst"


df.loc[df["title"].str.contains("engineer", case=False), "title"] = "Data_Engineer"
df.loc[df["title"].str.contains("ingeniero", case=False), "title"] = "Data_Engineer"
df.loc[df["title"].str.contains("data architect", case=False), "title"] = "Data_Engineer"
df.loc[df["title"].str.contains("programador", case=False), "title"] = "Data_Engineer"
df.loc[df["title"].str.contains("desarrolladores", case=False), "title"] = "Data_Engineer"
df.loc[df["title"].str.contains("developer", case=False), "title"] = "Data_Engineer"
df.loc[df["title"].str.contains("desarrollador", case=False), "title"] = "Data_Engineer"


df.loc[df["title"].str.contains("scientist", case=False), "title"] = "Data_Scientist"
df.loc[df["title"].str.contains("cientifico", case=False), "title"] = "Data_Scientist"
df.loc[df["title"].str.contains("research", case=False), "title"] = "Data_Scientist"


df.loc[df["title"].str.contains("business", case=False), "title"] = "Business_Analyst"

#Checking title columnn
df.title.value_counts()

#Checking title columnn
df.company.value_counts()

df.company=df.company.str.lstrip()
df.company=df.company.str.rstrip()
df.company=df.company.str.lower()

#Export data to CSV
df=df.to_csv('df.csv',encoding='utf-8-sig', index=False)
